create pytest conftest.py for relevant subfolders
