+++ tests/metabolomics/test_metabolomics.py
@@ -1,27 +1,7 @@
-import pytest
 from nplinker.metabolomics.metabolomics import make_families
 from nplinker.metabolomics.metabolomics import load_dataset
-from nplinker.metabolomics.metabolomics import load_spectra
-from nplinker.metabolomics.molecular_family import MolecularFamily
-from nplinker.metabolomics.spectrum import Spectrum
 from nplinker.strain_collection import StrainCollection
 from .. import DATA_DIR
-
-
-@pytest.fixture
-def spec_dict() -> dict[int, Spectrum]:
-    mgf_file = DATA_DIR / "spectra.mgf"
-    edges_file = DATA_DIR / "edges.pairsinfo"
-    return load_spectra(mgf_file, edges_file)
-
-@pytest.fixture
-def spec_with_families(spec_dict) -> dict[int, Spectrum]:
-    make_families(spec_dict.values())
-    return spec_dict
-
-@pytest.fixture
-def molecular_families(spec_dict) -> list[MolecularFamily]:
-    return make_families(spec_dict.values())
 
 
 def test_load_spectra(spec_dict):
