+++ tests/test_strain_collection.py
@@ -1,9 +1,6 @@
 import pytest
 from nplinker.strain_collection import StrainCollection
 from nplinker.strains import Strain
-from tests import DATA_DIR
-
-from .test_strain import strain
 
 
 @pytest.fixture
@@ -12,13 +9,6 @@ def collection(strain: Strain) -> StrainCollection:
     sut.add(strain)
     return sut
 
-
-@pytest.fixture
-def collection_from_file() -> StrainCollection:
-    filename = DATA_DIR / "strain_mappings.csv"
-    sut = StrainCollection()
-    sut.add_from_file(filename)
-    return sut
 
 def test_default():
     sut = StrainCollection()
@@ -73,7 +63,7 @@ def test_remove(collection: StrainCollection, strain: Strain):
 
     with pytest.raises(KeyError):
         collection.lookup(strain.id)
-    
+
     assert strain not in collection
 
     # needs fixing, see #90
